Add navigation dots
Added navigation dots for the slides.
Each circle represents a slide, so whenever a new
slide is activated, its corresponding circle gets
filled in so you can tell where in the show you are.
Each circle is linked to that particular slide,
so you can click on the circle and it will jump to that slide.